refactor: querytypes.Context as clienttypes.QueryContext
